Sets _force to True when parsing schemas (#31)
Since the serializer's encode with schema method causes
a register of the schema, the schema in the registry will
be updated with the special fastavro hint to signify that
fastavro has already parsed the schema. This can be problematic
in a distributed system since other Python fastavro consumers
will not actually have parsed the schema. This can trigger
a KeyError in fastavro because fastavro maintains a module
level dictionary of all the schemas it has parsed.